Tarea #19924: Cambio de atributo peso de integer a double en Acción. También en los lugares donde se lo utiliza.

+++ src/main/java/py/gov/stp/tools2/SqlInserts.java
@@ -293,7 +293,7 @@ public class SqlInserts {
 		
 		
 		insert.setDouble (1, accion.getCosto());
-		insert.setInt (2, accion.getPeso());
+		insert.setDouble(2, accion.getPeso());
 		//insert.setDate (3, sqlStartDate);
 		//insert.setDate (4, sqlEndDate);
 		insert.setInt(3, accion.getVersion());
